Fix out of bounds write issue
The inner loop had `size = min(len, size);` as size condition.
`len` is the total amount of bytes to write and thus isn't
valid after the first write.
The result is stack corruption crashing the userspace application.

This also adds some debug info to have a bit more info in case
of issues.